% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie_endpoints.R
\name{cookie_endpoint_simple}
\alias{cookie_endpoint_simple}
\title{Source code for a simple cookie endpoint}
\usage{
cookie_endpoint_simple(default_cookie_name = "therealshinyapps")
}
\arguments{
\item{default_cookie_name}{Character. The name of the cookie to use if the
user does not request a different cookie.}
}
\value{
The source code for a simple cookie endpoint as character.
}
\description{
Posit Cloud uses a cookie to identify the user. The output of this function
can be used to add a "/cookie" endpoint to an api in order to allow users to
easily extract their Posit Cloud cookie for your endpoint.
}
\examples{
cookie_endpoint_simple()
cookie_endpoint_simple("another_cookie_name")
}
