% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_pkg_api.R
\name{deploy_pkg_api}
\alias{deploy_pkg_api}
\title{Deploy a Package's API to Posit}
\usage{
deploy_pkg_api(
  package,
  folder_name,
  ...,
  cookie_endpoint_src = cookie_endpoint_httr2(),
  server = "posit.cloud",
  api_name = folder_name
)
}
\arguments{
\item{package}{Character. The name of the package that contains the API.}

\item{folder_name}{Character. The name of the folder that contains the API.}

\item{...}{Additional arguments to \code{\link[rsconnect:deployApp]{rsconnect::deployApp()}}.}

\item{cookie_endpoint_src}{Character. The source code for the cookie
endpoint, such as those returned by \code{\link[=cookie_endpoint_httr2]{cookie_endpoint_httr2()}} (the default)
and \code{\link[=cookie_endpoint_simple]{cookie_endpoint_simple()}}.}

\item{server}{Character. The server to deploy to. Defaults to "posit.cloud".}

\item{api_name}{Character. The name of the API. Defaults to \code{folder_name}.}
}
\value{
Whether the deployment was successful, invisibly.
}
\description{
\code{\link[plumber:plumb_api]{plumber::plumb_api()}} was added in plumber 1.0.0 to standardize the process
to wrap an API inside a package. However, deploying that API takes some extra
steps. This function streamlines that process.
}
