% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber2posit-package.R
\docType{package}
\name{plumber2posit-package}
\alias{plumber2posit}
\alias{plumber2posit-package}
\title{plumber2posit: Plumber on Posit Services}
\description{
Deploying APIs wrapped in packages can be tricky, and accessing APIs on 'Posit Cloud' is tricky. This package seeks to alleviate pain points in that process.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/plumber2posit}
  \item \url{https://jonthegeek.github.io/plumber2posit/}
  \item Report bugs at \url{https://github.com/jonthegeek/plumber2posit/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
