% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_pkg_api.R
\name{redeploy_pkg_api}
\alias{redeploy_pkg_api}
\title{Redeploy a Package's API to Posit}
\usage{
redeploy_pkg_api(
  package,
  folder_name,
  app_id,
  ...,
  cookie_endpoint_src = NULL,
  server = "posit.cloud"
)
}
\arguments{
\item{package}{Character. The name of the package that contains the API.}

\item{folder_name}{Character. The name of the folder that contains the API.}

\item{app_id}{Character or numeric. The number that identifies your API. You
can find this number in the URL of your API's page on Posit Cloud. For
example, for the app at \code{posit.cloud/content/12345}, the id is "12345".}

\item{...}{Additional arguments to \code{\link[rsconnect:deployApp]{rsconnect::deployApp()}}.}

\item{cookie_endpoint_src}{Character. The source code for the cookie
endpoint, such as those returned by \code{\link[=cookie_endpoint_httr2]{cookie_endpoint_httr2()}} and
\code{\link[=cookie_endpoint_simple]{cookie_endpoint_simple()}}. By default this is \code{NULL} here to remove the
cookie endpoint.}

\item{server}{Character. The server to deploy to. Defaults to "posit.cloud".}
}
\value{
Whether the deployment was successful, invisibly.
}
\description{
Once an API has been deployed to Posit Cloud, deploying something with the
same name will \emph{not} automatically update the existing API. Instead, you must
supply the \code{appId} argument to \code{\link[rsconnect:deployApp]{rsconnect::deployApp()}}, and \emph{not} the
\code{appName} argument. This function walks you through that process.
}
